+++ tasks/first_neighbor.py
@@ -1,4 +1,5 @@
-from drugstone.util.query_db import calculate_properties, map_edges, name2index, query_proteins_by_identifier
+from drugstone.util.property_calulations import calculate_properties, name2index
+from drugstone.util.query_db import map_edges, query_proteins_by_identifier
 from tasks.util.custom_network import add_edges, remove_ppi_edges
 from tasks.task_hook import TaskHook
 import graph_tool as gt
