+++ tasks/leiden_clustering.py
@@ -1,4 +1,4 @@
-from drugstone.util.query_db import calculate_properties
+from drugstone.util.property_calulations import calculate_properties
 from tasks.util.custom_network import add_edges, remove_ppi_edges
 from tasks.task_hook import TaskHook
 import graph_tool as gt
